Fix anyrun websearch
Couldn't find it anywhere in the documentations, but anyrun's websearch plugin relies on the XDG Mime Apps program. In particular, it executes a web search via 'xdg-open', which interestingly enough, is not added by setting xdg.enable and xdg.mimeApps.enable to true. It is therefore necessary to add the xdg-utils package to the list of anyrun's dependency.