Fix issues regarding sops executing too early/age keys mounted too late.
Sops decrypts secrets very early on in the boot process (stage 2). By then, the home dataset was not mounted yet, and therefore it would fail to find the keys. The only other dataset that was mounted early enough was /nix/persist (the actual persisted directories weren't bindmounted early enough either), and so the keys needed to be moved there.